fix: resolve clippy warnings and allow admin bypass on branch protection
Clippy fixes:
- Remove unnecessary u64 cast in config override
- Collapse nested if into match guard in event handler
- Collapse nested if statements in permission check
- Add Default implementation for EventDispatcher

+++ src/config/settings.rs
@@ -91,7 +91,7 @@ impl Settings {
             builder = builder.set_override("discord.token", token)?;
         }
         if let Some(app_id) = cli.application_id {
-            builder = builder.set_override("discord.application_id", app_id as u64)?;
+            builder = builder.set_override("discord.application_id", app_id)?;
         }
         if let Some(port) = cli.webhook_port {
             builder = builder.set_override("webhook.bind_port", port)?;
